Added Undefined component

+++ src/app/store/effects/project.effects.ts
@@ -62,7 +62,7 @@ export class ProjectEffects {
 		switchMap(([project, selectedProject]) => {
 			project.slug = project.title.replace(/ /g, ".");
 			if (selectedProject) {
-				project.id = selectedProject.id
+				project.id = selectedProject.id;
 				this._projectService.updateProject(project);
 			} else {
 				this._projectService.addProject(project);
@@ -78,17 +78,18 @@ export class ProjectEffects {
 	deleteProject$ = this._actions$.pipe(
 		ofType<DeleteProject>(EProjectActions.DeleteProject),
 		map(action => action.payload),
-		withLatestFrom(this._store.pipe(select(selectedProject))),
-		switchMap(([project, selectedProject]) => {
-			project.slug = project.title.replace(/ /g, ".");
-			if (selectedProject) {
-				project.id = selectedProject.id
+		withLatestFrom(this._store.pipe(select(selectLoggedInUserUID))),
+		switchMap(([project, selectLoggedInUserUID]) => {
+			if (selectLoggedInUserUID == project.userID) {
+				this._projectService.deleteProject(project);
+				return [
+					new SetSuccessMsg("Project Deleted Successfully!"),
+					new DeleteProjectSuccess(),
+					new NavigateToRoute(["home"])
+				];
+			} else {
+				 new SetErrorMsg("You don't have permission to delete this project")
 			}
-			this._projectService.deleteProject(project);
-			return [
-				new SetSuccessMsg("Project Deleted Successfully!"),
-				new DeleteProjectSuccess()
-			];
 		}),
 		catchError(err => of(new SetErrorMsg(err)))
 	);
@@ -132,7 +133,7 @@ export class ProjectEffects {
 		})
 	);
 
-	@Effect({dispatch: false})
+	@Effect({ dispatch: false })
 	getUserProjects$ = this._actions$.pipe(
 		ofType<GetUserProfileSuccess>(EAuthActions.GetUserProfileSuccess),
 		withLatestFrom(this._store.pipe(select(selectLoggedInUser))),
@@ -145,7 +146,7 @@ export class ProjectEffects {
 		})
 	);
 
-  /* 
+	/* 
   This effect handles logic between creating a new project 
   or editing an viewing and editing an existing project 
   */
@@ -162,11 +163,13 @@ export class ProjectEffects {
 					if (project) {
 						return of(new GetSelectedProjectFromRouteSuccess(project));
 					} else {
-						return of(new NewProject({
-              title: "NEW PROJECT",
-              type: action.payload,
-              user: "NOT YET ASSIGNED"
-            }));
+						return of(
+							new NewProject({
+								title: "NEW PROJECT",
+								type: action.payload,
+								userID: "NOT YET ASSIGNED"
+							})
+						);
 					}
 				}),
 				catchError(err => {
@@ -176,7 +179,7 @@ export class ProjectEffects {
 		})
 	);
 
-  /* 
+	/* 
   On saving a new project
   Update project name
   */
@@ -187,7 +190,7 @@ export class ProjectEffects {
 		switchMap((project: IProject) => {
 			this._userService.updateProjectName(project);
 			return [
-				new UpdateUserProfileSuccess(project.user),
+				new UpdateUserProfileSuccess(project.userID),
 				new NavigateToRoute([project.profile])
 			];
 		})
